Struct libra_security_logger::SecurityLog

#[must_use = "must use `log()`"]
pub struct SecurityLog { /* fields omitted */ }

The SecurityLog struct is used to log security-sensitive operations, for instance when an invalid signature is detected or when an unexpected event happens.

The security_log() function should be used to instantiate this struct. It can be decorated with different type of metadata:

event contains a pre-defined element in the SecurityEvent enum

error can contain an error type provided by the application

data can contain associated metadata related to the event

backtrace can contain a backtrace of the current call stack

All these information can be defined by using the appropriate function with the same name.

The method log() needs to be called to ensure that the event is actually printed.

Example:

use libra_security_logger::{security_log, SecurityEvent};
use std::fmt::Debug;

#[derive(Debug)]
struct SampleData {
    i: u8,
    s: Vec<u8>,
}

#[derive(Debug)]
enum TestError {
    Error,
}

    security_log(SecurityEvent::InvalidTransactionAC)
        .error(&TestError::Error)
        .data(&SampleData {
            i: 0xff,
            s: vec![0x90, 0xcd, 0x80],
        })
        .data("additional payload")
        .backtrace(100)
        .log();

In this example, security_log() logs an event of type SecurityEvent::InvalidTransactionAC, having TestError::Error as application error, a SimpleData struct and a String as additional metadata, and a backtrace that samples 100% of the times.

Methods

impl SecurityLog

pub fn data<T: Debug>(self, data: T) -> Self

Adds additional metadata to the SecurityLog struct. The argument needs to implement the std::fmt::Debug trait.

pub fn error<T: Debug>(self, error: T) -> Self

Adds an application error to the SecurityLog struct. The argument needs to implement the std::fmt::Debug trait.

pub fn backtrace(self, sampling_rate: u8) -> Self

Adds a backtrace to the SecurityLog struct.

pub fn log(self)

Prints the SecurityEvent struct.